Refactor Widgets
